feat: make match function pluggable
i.e. allow users to send in their match functions or use the in-built ones

This API change is mostly done to have better tree-shaking
